requirements.txt
django
daphne
channels
channels[daphne]
Pillow